apps/backend/pyproject.toml
[project]
name = "alpha-hub-backend"
version = "0.1.0"
description = "Alpha IT Hub - SaaS Backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.133.1",
    "httpx>=0.28.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.47",
    "structlog>=25.5.0",
    "supabase>=2.28.0",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
